Почти закончил классы модели.

+++ ProductionLineApp/Statistics.h
@@ -2,11 +2,11 @@
 
 struct Statistics
 {
-	float fstWorkplaceLoad;	// процент загрузки рабочих мест
-	float sndWorkplaceLoad;
-	float avgProdProcTime;	// в секундах (после запятой проценты миллисекунд, например 2,17 секунд)
+	unsigned long fstWorkplaceLoad;	// процент загрузки рабочих мест
+	unsigned long sndWorkplaceLoad;
+	unsigned long avgProdProcTime;	// в секундах (после запятой проценты миллисекунд, например 2,17 секунд)
 	int deferredProductsCount;
 	int fstWorkplaceAvgQueue;
 	int sndWorkplaceAvgQueue;
-	float fstWorkplaceBlockingTime;	// в секундах (после запятой проценты миллисекунд, например 2,17 секунд)
+	unsigned long fstWorkplaceBlockingTime;	// в секундах (после запятой проценты миллисекунд, например 2,17 секунд)
 };
